# allow local + grouped reduce in hand_coded