I tried to test it on all hw I have. gpt2 gets the most noticeable boost, while the other models remained relatively stable with no significant performance drop or gain.

### 7900xtx (hip)

| **Model**         | **Master** | **New**   | **Improvement (%)** |
|:-----------------:|:----------:|:---------:|:-------------------:|
| **GPT2 total**    | 10.75 ms   | 7.62 ms   | 29.2% ↑             |
| **LLAMA**         | N/A        | N/A       | N/A                 |
| **SD step**       | 367.60 ms  | 368.48 ms | 0.24% ↓             |
| **HLB CIFAR run** | 15.63 ms   | 15.48 ms  | 0.96% ↑             |

### 3090 (cuda)

| **Model**         | **Master** | **New**   | **Improvement (%)** |
|:-----------------:|:----------:|:---------:|:-------------------:|
| **GPT2 total**    | 11.91 ms   | 10.83 ms  | 9.08% ↑             |
| **LLAMA total**   | 41.52 ms   | 41.64 ms  | 0.29% ↓             |
| **SD step**       | 579.83 ms  | 571.83 ms | 1.38% ↑             |
| **HLB CIFAR run** | 31.56 ms   | 31.74 ms  | 0.57% ↓             |

### 4070ti (cuda)

| **Model**         | **Master** | **New**   | **Improvement (%)** |
|:-----------------:|:----------:|:---------:|:-------------------:|
| **GPT2 total**    | 12.66 ms   | 9.35 ms   | 26.18% ↑            |
| **SD step**       | 519.08 ms  | 518.16 ms | 0.18% ↑             |
| **HLB CIFAR run** | 30.01 ms   | 29.98 ms  | 0.1% ↑              |

### M1pro (metal)

| **Model**         | **Master** | **New**   | **Improvement (%)** |
|:-----------------:|:----------:|:---------:|:-------------------:|
| **GPT2 total**    | 45.39 ms   | 20.71 ms  | 54.39% ↑            |
| **LLAMA total**   | 138.40 ms  | 138.86 ms | 0.33% ↓             |
| **SD total**      | 1618.79 ms | 1617.58 ms| 0.07% ↑             |
| **HLB CIFAR run** | 202.6 ms   | 202.82 ms | 0.11% ↓             |