@geohot 2023-10-06
Seems fine. We really have to get the search working better