@geohot 2023-10-06
Hmm, this broke my model. OpenCL can't have local size product more than 256, and it's [4, 128, 1] here

[128, 1, 1] [4, 128, 1]
Traceback (most recent call last):
  File "/Users/kafka/fun/tinygrad/openpilot/compile2.py", line 69, in <module>
    run_schedule(schedule)
  File "/Users/kafka/fun/tinygrad/tinygrad/realize.py", line 62, in run_schedule
    out.realized = Device[out.device].exec_ast(op, output=out, inputs=buffers, var_vals=out.var_vals, **out._device_extra_args())
  File "/Users/kafka/fun/tinygrad/tinygrad/ops.py", line 259, in exec_ast
    prg.exec(rawbuffers, var_vals=var_vals)
  File "/Users/kafka/fun/tinygrad/tinygrad/ops.py", line 182, in exec
    return self(rawbufs, var_vals, force_wait=force_wait)
  File "/Users/kafka/fun/tinygrad/tinygrad/ops.py", line 193, in __call__
    if et := self.clprg(global_size, local_size, *rawbufs, *var_vals.values(), wait=force_wait or DEBUG>=1): GlobalCounters.time_sum_s += et
  File "/Users/kafka/fun/tinygrad/tinygrad/runtime/ops_gpu.py", line 97, in __call__
    e = self.clprgs[cl_bufs[0].device](CL.cl_queue[cl_bufs[0].device], [g*l for g,l in zip(global_size, local_size)] if local_size is not None else global_size, local_size, *cl_bufs, wait_for=wait_for)
  File "<pyopencl invoker for 'r_128_128_4_4_4'>", line 12, in enqueue_knl_r_128_128_4_4_4
pyopencl._cl.LogicError: clEnqueueNDRangeKernel failed: INVALID_WORK_GROUP_SIZE