@geohot 2023-10-06
The  # are we upcasting in mid reduce? (only for images) is triggering, adding a 4 to a 128