@geohot 2023-10-06
Also tanks performance on comma 3 device. I don't think we can merge any more changes to the hand coded optimization, reverted.